ParIterResult renamed as ParIterFallible
