+++ src/computational_variants/map.rs
@@ -1,7 +1,7 @@
 use super::xap::ParXap;
-use crate::ParIterResult;
+use crate::ParIterFallible;
 use crate::computational_variants::result::ParMapFallible;
-use crate::par_iter_result::IntoResult;
+use crate::par_iter_fallible::IntoResult;
 use crate::values::{Vector, WhilstAtom};
 use crate::{
     ChunkSize, IterationOrder, NumThreads, ParCollectInto, ParIter, ParIterUsing, Params,
@@ -178,7 +178,7 @@ where
         ParXap::new(params, iter, x1)
     }
 
-    fn into_fallible<Out, Err>(self) -> impl ParIterResult<R, Success = Out, Error = Err>
+    fn into_fallible<Out, Err>(self) -> impl ParIterFallible<R, Success = Out, Error = Err>
     where
         Self::Item: IntoResult<Out, Err>,
         Err: Send,
