+++ src/computational_variants/result/map_fallible.rs
@@ -1,6 +1,6 @@
 use crate::computational_variants::ParMap;
 use crate::computations::X;
-use crate::par_iter_result::{IntoResult, ParIterResult};
+use crate::par_iter_fallible::{IntoResult, ParIterFallible};
 use crate::runner::{DefaultRunner, ParallelRunner};
 use crate::{IterationOrder, ParCollectInto, ParIter};
 use orx_concurrent_iter::ConcurrentIter;
@@ -34,7 +34,7 @@ where
     }
 }
 
-impl<I, T, E, O, M1, R> ParIterResult<R> for ParMapFallible<I, T, E, O, M1, R>
+impl<I, T, E, O, M1, R> ParIterFallible<R> for ParMapFallible<I, T, E, O, M1, R>
 where
     R: ParallelRunner,
     I: ConcurrentIter,
@@ -52,7 +52,7 @@ where
 
     // computation transformations
 
-    fn map<Out, Map>(self, map: Map) -> impl ParIterResult<R, Success = Out, Error = Self::Error>
+    fn map<Out, Map>(self, map: Map) -> impl ParIterFallible<R, Success = Out, Error = Self::Error>
     where
         Map: Fn(Self::Success) -> Out + Sync,
         Out: Send,
