+++ src/computational_variants/xap.rs
@@ -1,6 +1,6 @@
-use crate::ParIterResult;
+use crate::ParIterFallible;
 use crate::computational_variants::result::ParXapFallible;
-use crate::par_iter_result::IntoResult;
+use crate::par_iter_fallible::IntoResult;
 use crate::values::TransformableValues;
 use crate::values::runner_results::Infallible;
 use crate::{
@@ -197,7 +197,7 @@ where
         ParXap::new(params, iter, x1)
     }
 
-    fn into_fallible<Out, Err>(self) -> impl ParIterResult<R, Success = Out, Error = Err>
+    fn into_fallible<Out, Err>(self) -> impl ParIterFallible<R, Success = Out, Error = Err>
     where
         Self::Item: IntoResult<Out, Err>,
         Err: Send,
