+++ src/lib.rs
@@ -21,8 +21,8 @@ mod into_par_iter;
 pub mod iter;
 mod iter_into_par_iter;
 mod par_iter;
+mod par_iter_fallible;
 mod par_iter_option;
-mod par_iter_result;
 mod parallel_drainable;
 mod parallelizable;
 mod parallelizable_collection;
@@ -51,8 +51,8 @@ pub use collect_into::ParCollectInto;
 pub use into_par_iter::IntoParIter;
 pub use iter_into_par_iter::IterIntoParIter;
 pub use par_iter::ParIter;
+pub use par_iter_fallible::ParIterFallible;
 pub use par_iter_option::ParIterOption;
-pub use par_iter_result::ParIterResult;
 pub use parallel_drainable::ParallelDrainableOverSlice;
 pub use parallelizable::Parallelizable;
 pub use parallelizable_collection::ParallelizableCollection;
